requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
boto3==1.26.54
botocore==1.29.54
Django==4.1.5
djangorestframework==3.14.0
gunicorn==20.1.0
jmespath==1.0.1
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
